Add sticker layer styles * fix layer panel styling
