+++ assets/src/edit-story/elements/sticker/layer.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,29 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/**
+ * External dependencies
+ */
+import { __ } from '@web-stories-wp/i18n';
 
 /**
  * Internal dependencies
  */
-import StoryPropTypes from '../../types';
-import stickers from '../../stickers';
+import { LayerText } from '../shared/layerText';
 
-const style = {
-  display: 'block',
-  height: 20,
-  width: 'auto',
-};
-
-const Noop = () => null;
-
-function StickerLayerContent({ element }) {
-  const { sticker } = element;
-  const Sticker = stickers[sticker.type]?.svg || Noop;
-  return <Sticker style={style} />;
+function StickerLayerContent() {
+  return <LayerText>{__('Sticker', 'web-stories')}</LayerText>;
 }
-
-StickerLayerContent.propTypes = {
-  element: StoryPropTypes.elements.sticker.isRequired,
-};
 
 export default StickerLayerContent;
